Fixes issue #3343 - Bump code coverage
